+++ CHANGELOG.md
@@ -200,6 +200,7 @@ Initial release.
     - Add `Error.IsSensitive` property. This property determines whether the `Error.ToString()` method results in a full representation of the error or an abbreviated representation.
     - Add `FailResult.SetCallbackFunction` and `FailResult.SetReplaceErrorFunction` methods. Functions passed to these methods are invoked whenver a `Fail` result is created.
     - Add `Result<T>.OnFailExcludingNone` methods.
+    - Add `AsNullable` result extension methods for `struct` types.
 
 ### Removed